server/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pafy==0.5.5
pycodestyle==2.9.1
python-vlc==3.0.16120
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
xmltodict==0.13.0
youtube-dl==2020.12.2
youtube-search==2.1.2
